Split Name::Parse tests out of name_test.rb (#4708)
name_test.rb was 4,456 lines - by far the largest model test file.
Name itself is already split into concern modules under
app/models/name/*.rb; this starts splitting the test file to match,
beginning with the biggest chunk: parsing.

Moved all 68 tests exercising Name::Parse (parse_name, parse_classification's
siblings excluded - those belong to Name::Taxonomy - standardize_name/
standardize_author/squeeze_author, the UPPER_WORD/LOWER_WORD/AUTHOR_PAT/
GROUP_PAT patterns, and the name_parse_* format-round-trip tests) plus
their private helpers (do_name_parse_test, assert_name_match*,
assert_name_parse_fails) into test/models/name/parse_test.rb as
Name::ParseTest, matching the Image::DhashTest/Observation::MergedNamingTest
naming convention already established elsewhere.

test_find_or_create_name_and_parents stayed behind despite living in the
old file's "Test name parsing" section - it exercises Name::Create's
find_or_create_name_and_parents, not Name::Parse.

name_test.rb now require_relative's the split file explicitly, so
`bin/rails test test/models/name_test.rb` on its own still runs all 183
tests (verified: same count before and after, and running both files
together doesn't double-count since require_relative is idempotent) -
this was the issue's explicit ask, since Rails' directory-based test
discovery alone doesn't pick up a split file when you target the main
file by name.
